refactor: refactor fetch video name to include service
