+++ src/services/index.js
@@ -1,1 +1,2 @@
-export { fetchVideos } from './fetch-videos';
+export { fetchVideos } from './fetch-videos.service';
+export { authService } from './auth.service';
